python/day2/requirements.txt
coverage==7.4.2
mysql-connector-python==8.3.0
